requirements.txt
praw
pytest
pyyaml
chevron